Extract controllers for /auth/* routes
