--- src/router/auth/validation.ts
@@ -1,9 +1,0 @@
-import RequestValidation, { Joi, Segments } from "../request-validation";
-
-export const login = new RequestValidation<object | undefined>({
-	[Segments.BODY]: Joi.object().optional(),
-});
-
-export const renew = new RequestValidation<object | undefined>({
-	[Segments.BODY]: Joi.object().optional(),
-});
